feat: banish the web server
The issue is thus: The "critical" feature to provide, is to have an `ipykernel` running inside of Blender. For convenience, we had decided to additionally bundle the `jupyter-server` webserver, as well as the `jupyterlab` GUI. The issue is one of simplicity and of dependencies: The newer version of `jupyterlab` that allows for use of existing kernels (without cursed provisioner mangling) unfortunately relies on a version of `requests` and/or `charset-normalizer` that conflicts with that vendored by Blender. So the decision was made to not bundle a webserver with this extension - _at all_.

BREAKING CHANGE: The extension no longer provides a web service. Please use the kernel connection file directly, or with an externally defined web service.