🐛 fix!: respect `bpy.app.online_access`
Starting an `ipykernel` opens up five sockets and ports. Arguably, this counts as "online access", since this may make it possible (if allowed by the host system) for networked users to connect to the running kernel.

For now, to be on the safe side, we make a simple `self.report()` of type `ERROR` if the user attempts to start the kernel without `bpy.app.online_access` being enabled. We ought to make a better UI, however.

BREAKING CHANGE: Use of the extension now requires enabling `bpy.app.online_access`.