✨ feat: comprehensive new UI
A new UI was added that makes it easy to get an overview of whether the kernel is running, as well as view and copy almost every piece of information about the kernel to the system clipboard.

This includes path to the connection file and its parent, multiline strings showing the absolute connection file path and its JSON contents (without printing the secret key), the IP and every port, the kernel key (copyable but with obfuscated printing) and transport mechanism, etc. .

Everything is contained in a mindful panel setup, which provides a rich explanation when online access is disabled (and what one can do about that).